% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cli_colors}
\alias{cli_colors}
\title{color of clinical varis}
\format{
A dataframe
}
\usage{
cli_colors
}
\value{
A dataframe
}
\description{
namesd vector, color of clinical varis
}
\keyword{datasets}
