% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virus_genome.R
\name{deal_virus_annotation}
\alias{deal_virus_annotation}
\title{Deal virus annotation}
\usage{
deal_virus_annotation(gene_features)
}
\arguments{
\item{gene_features}{result of get_virus_annotation()}
}
\value{
data.frame
}
\description{
Deal virus annotation
}
\examples{
gene_features <- get_virus_annotation(accession_number = "NC_001526.2",
    email = "13766876214@163.com")
virus_info <- deal_virus_annotation(gene_features)
}
