% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot_gene.R
\name{get_hot_gene}
\alias{get_hot_gene}
\title{Get hot insert genes by Chi-square test}
\usage{
get_hot_gene(virus_info, insert_info, tssRegion = c(-3000, 3000))
}
\arguments{
\item{virus_info}{The virus information data frame contains at least three
columns.
The first column is the gene name, the second column is the start site,
and the third column is the stop site.}

\item{insert_info}{The virus insertion information data frame contains at
least four columns.
The first column is the chromosome,
the second column is the host insertion site,
the third column is the virus break site, and the fourth column is the
number of reads.}

\item{tssRegion}{Region Range of TSS}
}
\value{
list
}
\description{
Get hot insert genes by Chi-square test
}
\examples{
data(insert_info)
virus_info <- data.frame(
      gene = c("E6", "E7", "E1", "E2", "E4", "E5", "L2", "L1", "LCR"),
      start = c(83, 562, 865, 2755, 3332, 3849, 4236, 5560, 7200),
      end = c(559, 858, 2813, 3852, 3619, 4100, 5657, 7155, 7904))
hot_gene <- get_hot_gene(virus_info, insert_info)
}
