% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virus_genome.R
\name{get_virus_genom}
\alias{get_virus_genom}
\title{Get virus genome}
\usage{
get_virus_genom(accession_number, email, output = NULL)
}
\arguments{
\item{accession_number, }{such as NC_001526.2.}

\item{email}{email address.}

\item{output}{Output file name, if not NULL, will be exported to the file.}
}
\value{
character
}
\description{
Get virus genome
}
\examples{
genome <- get_virus_genom(accession_number = "NC_001526.2",
    email = "13766876214@163.com")
}
