% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{insert_info}
\alias{insert_info}
\title{Insert information}
\format{
A dataframe
}
\usage{
insert_info
}
\value{
A dataframe
}
\description{
The virus insertion information data frame contains four columns.
The first column is the chromosome, the second column is the host insertion site,
the third column is the virus break site, and the fourth column is the number of reads.
}
\keyword{datasets}
