% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoplot.R
\name{oncoplot_clinical}
\alias{oncoplot_clinical}
\title{oncoplot_clinical}
\usage{
oncoplot_clinical(
  clinical_data,
  mapping,
  clinical_color = NULL,
  p_main = p_main
)
}
\arguments{
\item{clinical_data}{dataframe of clinical data,
the first column is sample ID.}

\item{mapping}{mapping.}

\item{clinical_color}{namesd vector, color of clinical varis.}

\item{p_main}{p_main.}
}
\value{
a gg object
}
\description{
oncoplot_clinical
}
