% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoplot.R
\name{oncoplot_fill}
\alias{oncoplot_fill}
\title{oncoplot_fill}
\usage{
oncoplot_fill(
  breaks = NULL,
  varis_color = NULL,
  name = NULL,
  na.value = "#bdbdbd"
)
}
\arguments{
\item{breaks}{breaks of scale_fill_manual}

\item{varis_color}{namesd vector, color of varis}

\item{name}{name of fill legend}

\item{na.value}{color of NA value.}
}
\value{
a gg object
}
\description{
oncoplot_fill
}
