% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoplot.R
\name{oncoplot_gene}
\alias{oncoplot_gene}
\title{oncoplot_gene}
\usage{
oncoplot_gene(mat, ylab = "percentage", varis_color)
}
\arguments{
\item{mat}{matrix, each row is a gene,
and each column is a sample}

\item{ylab}{ylab}

\item{varis_color}{namesd vector, color of varis}
}
\value{
a gg object
}
\description{
oncoplot_gene
}
