% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoplot.R
\name{oncoplot_sample}
\alias{oncoplot_sample}
\title{oncoplot_sample}
\usage{
oncoplot_sample(mat, varis_color)
}
\arguments{
\item{mat}{matrix, each row is a gene,
and each column is a sample}

\item{varis_color}{namesd vector, color of varis}
}
\value{
a gg object
}
\description{
oncoplot_sample
}
