% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pdata}
\alias{pdata}
\title{clinical data}
\format{
A dataframe
}
\usage{
pdata
}
\value{
A dataframe
}
\description{
dataframe of clinical data, the first column is sample ID.
}
\keyword{datasets}
