% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vcf_matrix}
\alias{vcf_matrix}
\title{matrix of vcf data}
\format{
A dataframe
}
\usage{
vcf_matrix
}
\value{
A dataframe
}
\description{
each row is a gene, and each column is a sample
}
\keyword{datasets}
