% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoplot.R
\name{oncoplot_main}
\alias{oncoplot_main}
\title{main plot}
\usage{
oncoplot_main(mat, varis_color, na.value = "#bdbdbd")
}
\arguments{
\item{mat}{matrix, each row is a gene,
and each column is a sample}

\item{varis_color}{namesd vector, color of varis}

\item{na.value}{color of na value.}
}
\value{
a gg object
}
\description{
main plot
}
