% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strudel_plot.R
\name{strudel_plot}
\alias{strudel_plot}
\title{Strudel plot of virus insert}
\usage{
strudel_plot(
  virus_info,
  insert_info,
  virus_color = "#EAFEFF",
  host_color = "#EAFEFF",
  label_virus = "HPV16",
  label_host = "Host",
  hot_gene = 5,
  size_gene = 6.5,
  size_label = 6.5,
  text_repel = TRUE,
  read_cutoff = 0,
  pvalue_cutoff = 0.05,
  sample_select = NULL
)
}
\arguments{
\item{virus_info}{The virus information data frame contains at least three
columns.
The first column is the gene name, the second column is the start site,
and the third column is the stop site.}

\item{insert_info}{The virus insertion information data frame contains at
least four columns.
The first column is the chromosome,
the second column is the host insertion site,
the third column is the virus break site, and the fourth column is the
number of reads.}

\item{virus_color}{color of virus rect.}

\item{host_color}{color of host rect.}

\item{label_virus}{label of virus.}

\item{label_host}{label of host.}

\item{hot_gene}{number of insert hot genes.}

\item{size_gene}{size of gene labels.}

\item{size_label}{size of label_virus and label_host.}

\item{text_repel}{If TRUE (the default), use geom_text_repel to make text 
labels repel away from each other and away from the data points.}

\item{read_cutoff}{cutoff of read number. The default value is 0.}

\item{pvalue_cutoff}{cutoff of pvalue. The default value is 0.05.}

\item{sample_select}{samples to use. The default value is NULL.}
}
\value{
a gg object
}
\description{
Strudel plot of virus insert
}
\examples{
data(insert_info)
virus_info <- data.frame(
      gene = c("E6", "E7", "E1", "E2", "E4", "E5", "L2", "L1", "LCR"),
      start = c(83, 562, 865, 2755, 3332, 3849, 4236, 5560, 7200),
      end = c(559, 858, 2813, 3852, 3619, 4100, 5657, 7155, 7904))
p <- strudel_plot(virus_info, insert_info)
}
